Swiping to change pages didn't always work
When the swipe was initiated on a button the swipe didn't register.
Fixed by moving the buttons inside the Swipable component
