--- components/Reader/Backdrop.js
@@ -1,25 +1,0 @@
-// @flow
-/**
- * Part of GDL gdl-frontend.
- * Copyright (C) 2017 GDL
- *
- * See LICENSE
- */
-import styled from 'react-emotion';
-import media from '../../style/media';
-
-// A grey backdrop that's only visible on tablets/desktops
-const Backdrop = styled.div`
-  bottom: 0;
-  left: 0;
-  right: 0;
-  top: 0;
-  position: fixed;
-  z-index: -2;
-  background-color: rgba(0, 0, 0, 0.7);
-  ${media.mobile`
-    display: none;
-  `};
-`;
-
-export default Backdrop;
